CDO: (MAJOR++) Combination of many commits from the master branch
- Main features: Optimization in the assembly process and the Hodge building
- Several fixes
- The user settings may be break with this bundle of commits

Here are the commits from the master branch

Optimizations:
-------------

CDO: (Major) Renaming and modification of the function prototype for time-dependent quantities
CDO: Add a new definition type for quantities with a time variation only
   This commit may break a user-defined setting with a function for the
   definition of the time step
CDO: (minor) Modify cs_quant_set_face_nvec() to be more efficient
CDO: Optimize the building of the cell mesh according to the mesh building flag
CDO: Move code related to v2v and f2f adjacencies into cs_cdo_connect
CDO: Better monitoring of the assembly process
Inlining cs_timer_add_diff() to get a cheaper monitoring
CDO: (Major) Modify the way assembly process is performed
   Case of scalar- and vector-valued equations
   Use a structure which is owned by each thread and allocate the buffers
   in order to call the assembly function one time per cell
CDO: (Major) Switch to function pointers to handle the assembly process
   Modify the structure cs_equation_assembly_buf_t
CDO: Re-organization of a part of the setup step to avoid cyclic includes
Add a function to set function pointers for the assembly step.
Add new function pointers:
 - without openMP
 - with openMP with a different OpenMP choice: critical or atomic
 The default behavior with OpenMP is set to critical

CDO: Better OpenMP directives for monitoring the performance of the assembly
(Major) Add a dedicated function in cs_matrix_assembler for CDO to be more efficient
CDO: Improved way of building the stiffness matrix in scalar-valued Vb schemes
CDO: (Major) Optimized the assembly process for CDO matrices

  Remove layers and add functions to be more specific to CDO schemes
  A dedicated structure and a function pointer have been added to
  handled different cases (scalar-valued/vector-valued or sequential/parallel
  atomic/critical OpenMP sections).
CDO: Add a specific version for creating/initializing a cellwise block 3x3 matrix
CDO: Several optimizations in Vb schemes

Minor commits:
-------------

(minor) Adding missing Doxygen documentation
Avoid a compilation warning with intel compilers
CDO: Better automatic checkings to assess the correct definition of the cell mesh in debug mode
CDO: (minor) Fix or add comments
CDO: (minor) Avoid a compilation warning
Update AUTHORS file
ALE: (minor) remove a debug message
CDO: (minor) Remove several compilation warnings
Minor modifications in aggregation. Case of MSR matrices and CS_GRID_COARSENING_SPD_MX
(minor) Comments, typos, style fixes
Fix several compilation warnings
Dump more details in agglomeration algorithms when the verbosity level is high

Fixes:
-----

CDO: Fix 0b22f568 (2019-02-21). Wrong argument type.
CDO: Fix e62bfe75 (2019-02-20). Set the number of cell faces in all cases.

Linear algebra:
--------------

Add the initial residual value when the verbosity value is greater than 1
CDO: Add a new discrete Hodge operator building in scalar-valued Fb schemes
Add an exit in the gmres when the max number of iterations is reached
CDO: Modify the default settings for the K-cycle as preconditioner

+++ src/cdo/cs_cdo_local.h
@@ -94,7 +94,7 @@ typedef struct {
 
   /* Temporary buffers (erase and updated several times during the system
      build */
-  short int    *ids;     /*!< local ids */
+  int          *ids;     /*!< local ids */
   double       *values;  /*!< local values */
   cs_real_3_t  *vectors; /*!< local 3-dimensional vectors */
 
@@ -369,7 +369,7 @@ cs_cell_mesh_get_f2v(short int                    f,
   /* Tag vertices belonging to the current face f */
   for (short int i = cm->f2e_idx[f]; i < cm->f2e_idx[f+1]; i++) {
 
-    const short int  shift_e = 2*cm->f2e_ids[i];
+    const int  shift_e = 2*cm->f2e_ids[i];
     v_ids[cm->e2v_ids[shift_e]] = 1;
     v_ids[cm->e2v_ids[shift_e+1]] = 1;
 
@@ -446,16 +446,17 @@ cs_cdo_local_finalize(void);
  * \param[in]   n_max_fbyc      max number of faces in a cell
  * \param[in]   n_blocks        number of blocks in a row/column
  * \param[in]   block_sizes     size of each block or NULL if n_blocks = 1
+ *                              Specific treatment n_blocks = 1.
  *
  * \return a pointer to a new allocated \ref cs_cell_sys_t structure
  */
 /*----------------------------------------------------------------------------*/
 
 cs_cell_sys_t *
-cs_cell_sys_create(int          n_max_dofbyc,
-                   int          n_max_fbyc,
-                   short int    n_blocks,
-                   short int   *block_sizes);
+cs_cell_sys_create(int    n_max_dofbyc,
+                   int    n_max_fbyc,
+                   int    n_blocks,
+                   int   *block_sizes);
 
 /*----------------------------------------------------------------------------*/
 /*!
